Refactor user model to use integer for Gender and Status properties; update related migrations and seed data accordingly.
