--- src/HiveSpace.UserService/HiveSpace.UserService.Api/HiveSpace.UserService.Api.http
@@ -1,6 +1,0 @@
-@HiveSpace.UserService.Api_HostAddress = http://localhost:5014
-
-GET {{HiveSpace.UserService.Api_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
