+++ src/HiveSpace.UserService/HiveSpace.UserService.Application/Services/AdminService.cs
@@ -2,7 +2,7 @@ using HiveSpace.Core.Contexts;
 using HiveSpace.UserService.Application.Models.Requests.Admin;
 using HiveSpace.UserService.Application.Models.Responses.Admin;
 using HiveSpace.UserService.Application.Interfaces;
-using HiveSpace.UserService.Application.Queries;
+using HiveSpace.UserService.Application.Interfaces.DataQueries;
 using HiveSpace.UserService.Domain.Aggregates.User;
 using HiveSpace.UserService.Domain.Services;
 using HiveSpace.UserService.Domain.Repositories;
