+++ src/HiveSpace.UserService/HiveSpace.UserService.Infrastructure/Data/UserDbContext.cs
@@ -23,7 +23,6 @@ public class UserDbContext : IdentityDbContext<ApplicationUser, IdentityRole<Gui
         builder.ApplyConfigurationsFromAssembly(typeof(UserDbContext).Assembly);
 
         // Configure table names following the Identity Service pattern
-        builder.Entity<ApplicationUser>().ToTable("users");
         builder.Entity<Address>().ToTable("addresses");
         builder.Entity<IdentityRole<Guid>>().ToTable("roles");
         builder.Entity<IdentityUserRole<Guid>>().ToTable("user_roles");
@@ -32,13 +31,5 @@ public class UserDbContext : IdentityDbContext<ApplicationUser, IdentityRole<Gui
         builder.Entity<IdentityRoleClaim<Guid>>().ToTable("role_claims");
         builder.Entity<IdentityUserToken<Guid>>().ToTable("user_tokens");
         builder.Entity<Store>().ToTable("stores");
-
-        // Explicitly map ApplicationUser -> IdentityUserRole using UserId (avoid shadow FK ApplicationUserId)
-        builder.Entity<ApplicationUser>()
-            .HasMany(u => u.UserRoles)
-            .WithOne() // IdentityUserRole<TKey> has no navigation back to user by default
-            .HasForeignKey(ur => ur.UserId)
-            .IsRequired()
-            .OnDelete(DeleteBehavior.Cascade);
     }
 }
