+++ src/HiveSpace.UserService/HiveSpace.UserService.Infrastructure/Identity/CustomProfileService.cs
@@ -75,7 +75,7 @@ public class CustomProfileService : IProfileService
         }
         
         // 3. Add user status information
-        claims.Add(new Claim("user_status", user.UserStatus));
+        claims.Add(new Claim("user_status", user.Status.ToString()));
         
         // 4. Add additional user information if available
         if (user.DateOfBirth.HasValue)
@@ -83,9 +83,9 @@ public class CustomProfileService : IProfileService
             claims.Add(new Claim("birthdate", user.DateOfBirth.Value.ToString("yyyy-MM-dd")));
         }
         
-        if (!string.IsNullOrEmpty(user.Gender))
+        if (user.Gender.HasValue)
         {
-            claims.Add(new Claim("gender", user.Gender));
+            claims.Add(new Claim("gender", user.Gender.Value.ToString()));
         }
 
         // Finally, issue the claims to the token.
@@ -103,6 +103,6 @@ public class CustomProfileService : IProfileService
         
         // A user is considered active if they exist, are not locked out, and have an active status.
         context.IsActive = user is not null 
-            && user.UserStatus == UserStatus.Active.ToString();
+            && user.Status == (int)UserStatus.Active;
     }
 }
